% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aperol.R
\name{drunk}
\alias{drunk}
\title{Drunk Praise}
\usage{
drunk(
  repeat_words = 1,
  repeat_times = 2,
  template = "You are ${adverb} ${adjective}!"
)
}
\arguments{
\item{repeat_words}{int scalar, the number of words that get repeated}

\item{repeat_times}{an integer-valued (non-negative) vector. If \code{length(repeat_times)} is \code{repeat_words}, gives the number of times each repeated word appears. If length 1, each repeated word appears that number of times. Otherwise, throws an error (see \code{\link[=rep]{rep()}}).}

\item{template}{character scalar, template for the (non-swapped) praise}
}
\description{
Generates praise with \code{\link[praise:praise]{praise::praise()}}, then repeats some words, and permutes them all.
}
\examples{
drunk()
drunk(2, 2:3)
drunk(2, 2, "You are ${creating} a ${adverb} ${adjective} ${rpackage}")

}
