% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aperol.R
\name{drunk}
\alias{drunk}
\title{Drunk Praise}
\usage{
drunk(
  repeat_words = 1,
  repeat_times = 2,
  template = "You are ${adverb} ${adjective}!"
)
}
\arguments{
\item{repeat_words}{int, number of words that get repeated}

\item{repeat_times}{int, number of times to repeat the repeated words}

\item{template}{character scalar, template for the (non-swapped) praise}
}
\description{
Generates praise with \code{\link[praise:praise]{praise::praise()}}, then repeats some words, and permutes them all.
}
\examples{
drunk()
drunk(2, 3, "You are ${creating} a ${adverb} ${adjective} ${rpackage}")

}
