% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spritz.R
\name{spritz}
\alias{spritz}
\title{Spritz}
\usage{
spritz(
  runs = 3,
  minutes = 10,
  template = "You are ${adverb} ${adjective}!",
  func = "devtools::check()"
)
}
\arguments{
\item{runs}{Numeric. Number of runs in time frame before praise}

\item{minutes}{Numeric. Number of minutes in time frame}

\item{template}{Character. Praise template}

\item{func}{Character. A function you expect to run multiple times}
}
\value{
Returns the output of func
}
\description{
A thin wrapper around a function (e.g devtools::check())
that provides affirmation after X calls in Y minutes. Praise
gets drunker as the number of actual runs exceeds the number of
runs in the time frame.
}
\examples{
\dontrun{

# default is to run devtools::check()
spritz()
## change rate depending on speed of devtools::check()
spritz(runs = 4, minutes = 15)



}


}
